MonteCarloMeasurements API

MonteCarloMeasurements.AccumulatedSeries

MonteCarloMeasurements.MonteCarloMeasurement

MonteCarloMeasurements.TimeSeries

MonteCarloMeasurements.name

MonteCarloMeasurements.push!

MonteCarloMeasurements.AccumulatedSeries — Type

AccumulatedSeries{T <: Number} <: MonteCarloMeasurement

A type of MonteCarloMeasurement that accumulates statistics for a given observable while performing an online binning analysis provided by OnlineLogBinning.jl.

MonteCarloMeasurements.MonteCarloMeasurement — Type

MonteCarloMeasurement

Abstract type that provides an interface for all MonteCarloMeasurements.

Warning

Required Interface Functions

The following functions must have methods defined for each new MonteCarloMeasurement.

push!: Move a new measurement into the MonteCarloMeasurement instance.

Note

Default Interface Functions

The following functions have a default method for any given MonteCarloMeasurement. * name: Define a name for a given MonteCarloMeasurement

MonteCarloMeasurements.TimeSeries — Type

TimeSeries{T <: Number} <: MonteCarloMeasurement

A type of MonteCarloMeasurement that stores the measurements in a datastream::Vector{T}. Statistics are not accumulated in an online way.

MonteCarloMeasurements.name — Method

name(meas::MonteCarloMeasurement)

Return the name defined by the input MonteCarloMeasurement.

julia> meas = TimeSeries("Walter White");

julia> name(meas)
"Walter White"

MonteCarloMeasurements.push! — Method

push!(meas::AccumulatedSeries, value)

push! a single value or many values into a [AccumulatedSeries] datastream.